Refactor rp2 module: Update type hints, add DMA class, and enhance PIO functionality
- Updated type hints and imports for better clarity and compatibility.
- Introduced the `DMA` class for managing DMA controller channels, including methods for configuration, control, and IRQ handling.
- Enhanced `PIO` and `StateMachine` classes with additional methods for program management and state machine control.
- Added `bootsel_button` function to read the state of the BOOTSEL button.
- Deprecated `country` function with overloads for getting and setting country codes.
- Improved documentation and comments throughout the module for better understanding and usage.
